Do not specify precision on numpy types

+++ src/lammpsio/snapshot.py
@@ -204,7 +204,7 @@ class Snapshot:
     @id.setter
     def id(self, value):
         if value is not None:
-            v = numpy.array(value, ndmin=1, copy=False, dtype=numpy.int32)
+            v = numpy.array(value, ndmin=1, copy=False, dtype=int)
             if v.shape != (self.N,):
                 raise TypeError("Ids must be a size N array")
             if not self.has_id():
@@ -228,17 +228,17 @@ class Snapshot:
     def position(self):
         """:class:`numpy.ndarray`: Positions."""
         if not self.has_position():
-            self._position = numpy.zeros((self.N, 3), dtype=numpy.float64)
+            self._position = numpy.zeros((self.N, 3), dtype=float)
         return self._position
 
     @position.setter
     def position(self, value):
         if value is not None:
-            v = numpy.array(value, ndmin=2, copy=False, dtype=numpy.float64)
+            v = numpy.array(value, ndmin=2, copy=False, dtype=float)
             if v.shape != (self.N, 3):
                 raise TypeError("Positions must be an Nx3 array")
             if not self.has_position():
-                self._position = numpy.zeros((self.N, 3), dtype=numpy.float64)
+                self._position = numpy.zeros((self.N, 3), dtype=float)
             numpy.copyto(self._position, v)
         else:
             self._position = None
@@ -258,17 +258,17 @@ class Snapshot:
     def image(self):
         """:class:`numpy.ndarray`: Images."""
         if not self.has_image():
-            self._image = numpy.zeros((self.N, 3), dtype=numpy.int32)
+            self._image = numpy.zeros((self.N, 3), dtype=int)
         return self._image
 
     @image.setter
     def image(self, value):
         if value is not None:
-            v = numpy.array(value, ndmin=2, copy=False, dtype=numpy.int32)
+            v = numpy.array(value, ndmin=2, copy=False, dtype=int)
             if v.shape != (self.N, 3):
                 raise TypeError("Images must be an Nx3 array")
             if not self.has_image():
-                self._image = numpy.zeros((self.N, 3), dtype=numpy.int32)
+                self._image = numpy.zeros((self.N, 3), dtype=int)
             numpy.copyto(self._image, v)
         else:
             self._image = None
@@ -288,17 +288,17 @@ class Snapshot:
     def velocity(self):
         """:class:`numpy.ndarray`: Velocities."""
         if not self.has_velocity():
-            self._velocity = numpy.zeros((self.N, 3), dtype=numpy.float64)
+            self._velocity = numpy.zeros((self.N, 3), dtype=float)
         return self._velocity
 
     @velocity.setter
     def velocity(self, value):
         if value is not None:
-            v = numpy.array(value, ndmin=2, copy=False, dtype=numpy.float64)
+            v = numpy.array(value, ndmin=2, copy=False, dtype=float)
             if v.shape != (self.N, 3):
                 raise TypeError("Velocities must be an Nx3 array")
             if not self.has_velocity():
-                self._velocity = numpy.zeros((self.N, 3), dtype=numpy.float64)
+                self._velocity = numpy.zeros((self.N, 3), dtype=float)
             numpy.copyto(self._velocity, v)
         else:
             self._velocity = None
@@ -318,17 +318,17 @@ class Snapshot:
     def molecule(self):
         """:class:`numpy.ndarray`: Molecule tags."""
         if not self.has_molecule():
-            self._molecule = numpy.zeros(self.N, dtype=numpy.int32)
+            self._molecule = numpy.zeros(self.N, dtype=int)
         return self._molecule
 
     @molecule.setter
     def molecule(self, value):
         if value is not None:
-            v = numpy.array(value, ndmin=1, copy=False, dtype=numpy.int32)
+            v = numpy.array(value, ndmin=1, copy=False, dtype=int)
             if v.shape != (self.N,):
                 raise TypeError("Molecules must be a size N array")
             if not self.has_molecule():
-                self._molecule = numpy.zeros(self.N, dtype=numpy.int32)
+                self._molecule = numpy.zeros(self.N, dtype=int)
             numpy.copyto(self._molecule, v)
         else:
             self._molecule = None
@@ -348,17 +348,17 @@ class Snapshot:
     def typeid(self):
         """:class:`numpy.ndarray`: Types."""
         if not self.has_typeid():
-            self._typeid = numpy.ones(self.N, dtype=numpy.int32)
+            self._typeid = numpy.ones(self.N, dtype=int)
         return self._typeid
 
     @typeid.setter
     def typeid(self, value):
         if value is not None:
-            v = numpy.array(value, ndmin=1, copy=False, dtype=numpy.int32)
+            v = numpy.array(value, ndmin=1, copy=False, dtype=int)
             if v.shape != (self.N,):
                 raise TypeError("Type must be a size N array")
             if not self.has_typeid():
-                self._typeid = numpy.ones(self.N, dtype=numpy.int32)
+                self._typeid = numpy.ones(self.N, dtype=int)
             numpy.copyto(self._typeid, v)
         else:
             self._typeid = None
@@ -378,17 +378,17 @@ class Snapshot:
     def charge(self):
         """:class:`numpy.ndarray`: Charges."""
         if not self.has_charge():
-            self._charge = numpy.zeros(self.N, dtype=numpy.float64)
+            self._charge = numpy.zeros(self.N, dtype=float)
         return self._charge
 
     @charge.setter
     def charge(self, value):
         if value is not None:
-            v = numpy.array(value, ndmin=1, copy=False, dtype=numpy.float64)
+            v = numpy.array(value, ndmin=1, copy=False, dtype=float)
             if v.shape != (self.N,):
                 raise TypeError("Charge must be a size N array")
             if not self.has_charge():
-                self._charge = numpy.zeros(self.N, dtype=numpy.float64)
+                self._charge = numpy.zeros(self.N, dtype=float)
             numpy.copyto(self._charge, v)
         else:
             self._charge = None
@@ -408,17 +408,17 @@ class Snapshot:
     def mass(self):
         """:class:`numpy.ndarray`: Masses."""
         if not self.has_mass():
-            self._mass = numpy.ones(self.N, dtype=numpy.float64)
+            self._mass = numpy.ones(self.N, dtype=float)
         return self._mass
 
     @mass.setter
     def mass(self, value):
         if value is not None:
-            v = numpy.array(value, ndmin=1, copy=False, dtype=numpy.float64)
+            v = numpy.array(value, ndmin=1, copy=False, dtype=float)
             if v.shape != (self.N,):
                 raise TypeError("Mass must be a size N array")
             if not self.has_mass():
-                self._mass = numpy.ones(self.N, dtype=numpy.float64)
+                self._mass = numpy.ones(self.N, dtype=float)
             numpy.copyto(self._mass, v)
         else:
             self._mass = None
